requirements.txt
ply >=3.4,<4.0
python-constraint >=1.2,<2.0
python-iptables >=0.5,<0.6
pytest >=2.6,<2.7
scapy ==2.2.0-dev
